+++ pkgs/google_cloud_auth/test/service_account_credentials_test.dart
@@ -104,6 +104,8 @@ void main() {
             info,
           );
 
+          expect(creds, isA<GoogleCredentials>());
+          expect(creds, isA<ServiceAccountSigner>());
           expect(
             creds.tokenUri,
             equals(Uri.https('oauth2.googleapis.com', '/token')),