+++ pkgs/google_cloud_auth/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.0-wip
 
+* Added `GoogleCredentials` base class for all credential types.
 * Support for key signing using `ComputeEngineCredentials`,
   `ServiceAccountCredentials`, and Application Default Credentials
   (`defaultCredentials`).